Move coniks package one level down
